+++ packages/robotscope-core/src/index.ts
@@ -12,6 +12,7 @@ export * from "./ros2/pointcloud2.js";
 export * from "./ros2/pose.js";
 export * from "./ros2/trajectory.js";
 export * from "./ros2/occupancy-grid.js";
+export * from "./ros2/lanelet-map-bin.js";
 export * from "./tf/transform-math.js";
 export * from "./storage/sidecar.js";
 export * from "./storage/topic-time-index.js";